@@ -9,8 +9,8 @@ import com.google.common.io.ByteStreams;
 
 import java.util.UUID;
 import java.util.logging.Level;
-import net.md_5.bungee.api.chat.TextComponent;
 
+import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.connection.PendingConnection;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.connection.Server;
